Updated with changes to boilerplate code
[ci skip]

I also changed the section about adding a style. Whilst the intent of this section was to show passing attribute values to the created VNode, `styles` is a poor example as direct style manipulation, whilst possible, is discouraged (as it bypasses the themeing system). I changed this section to set a `title` on the node instead.