Enhance codefile with title and note (#446)
* Fix issue 443: code highlighting and code file names

Added ability to highlight specific lines of code embedded via include_codefile. Also added ability to include description/footer for the code block.

* Updated Theme tutorial to use codefile updates

Highlighted important lines in the file and added a file path to the code block description. This would need to be done to all the tutorials eventually, but this one is the proof of concept.

* Added desc to documentation

* Manually set lang

CSS wasn't being autodetected for *.m.css blocks.

* Added title and replaced desc with note

After trying to use the tutorial it was evident that having something at the top of the block was generally more useful. The code for the bottom already existed so I migrated `desc` to `note` (footnote) and created a new `title` attribute.